requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
flask==2.3.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.3.2
tensorflow==2.17.0
pymongo==4.4.0
python-dotenv==1.0.0
streamlit==1.27.1
matplotlib==3.8.1
plotly==5.22.0
scipy==1.11.3
requests==2.31.0
